part refacto
